backend/requirements.txt

fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.4
python-multipart==0.0.9
